Segment cells with Julia

Segment cells with Julia

Welcome to SegmentationTools.jl! This is a random collection of functions that I've implemented for segmenting single cells from microscopy images.

Flow

The general procedure follows the following format:

Segmentation -> Linking -> Analysis

Segmentation

SegmentationTools.segment_cells — Function.

segment_cells(img, seed_channel, segment_channel, maskfunc)

source

Linking

SegmentationTools.build_tp_df — Function.

build_tp_df(img, segments, signal_channel)

Build a DataFrames.DataFrame that is compatible with trackpys link_df function. Needs to be converted to a Pandas.DataFrame before passing

source

I generally use it like so:

particles = SegmentationTools.build_tp_df(img1, segments, :EPI_mNG);
using PyCall
tp = pyimport("trackpy")

import Pandas
# convert particles to a Pandas DataFrame and call trackpy
t = tp.link_df(Pandas.DataFrame(particles), 5);
linked = DataFrame(Pandas.DataFrame(t)); # convert back to Julian dataframe

Analysis and Plotting

SegmentationTools.create_cell_grid — Function.

r createcellgrid(img, tracks; win) -> grid

Given a xyt image and set of tracks, creates a grid of all tracked cells with the cells centered in their grid location. Helps to quickly diagnose weird cell segmentation behavior.

source

Miscellaneous

SegmentationTools.flatfield_correct — Function.

flatfield_correct(img, axis, darkfield, flatfield)

Flatfield correct img. This function subtracts the darkfield, clamps the output to positive values, divides by the flatfield, and then rescales and rebuilds the original image perserving the properties and axes.

For example, to correct along the DAPI channel:

flatfield_correct(img[Axis{:channel}(:EPI_DAPI)], darkfield, flatfield_dapi)

source

SegmentationTools.get_background_means — Function.

get_background_means(img, seeds, dist)

Computes the average signal of the background area, which is defined as areas in img that are a distance dist away from the true values in seeds for each slice in time.

Example:

a = img[Axis{:position}(2), Axis{:channel}(:EPImNG)].data b = img[Axis{:position}(2), Axis{:channel}(:EPIBFP)].data .> 0.01 getbackgroundmeans(a, b)

source