Neuron Segmentation API

Semantic Segmentation

Missing docstring.

Missing docstring for make_unet_input_h5, call_unet. Check Documenter's build log for details.

Watershed-Threshold Instance Segmentation

Missing docstring.

Missing docstring for instance_segmentation_watershed, instance_segmentation_threshold, detect_incorrect_merges, watershed_threshold, instance_segmentation, consolidate_labeled_img. Check Documenter's build log for details.

Watershed-Concave Instance Segmentation (currently not used)

SegmentationTools.instance_segment_concave — Function

Recursively segments all concave neurons in an image.

Arguments

img_roi: Image to segment

Optional keyword arguments

threshold_scale::Real: Neurons less concave than this won't be segmented. Default 0.3

num_neurons::Real: Maximum number of concave neurons per frame. Defaul 10.

zscale::Real: Scale of z-axis relative to xy plane. Default 1.

min_neuron_size::Integer: Minimum size of a neuron (in pixels). Default 10.

scale_recurse_multiply::Real: Factor to increase the concavity threshold for recursive segmentation. Default 1.5.

init_scale::Real: Amount to expand first neuron before computing location of second neuron. Default 0.7.