UNet Training API

SegmentationTools.create_weights — Function

Creates a weight map from a given labeled dataset. Unlabeled data (label 0) has weight 0 and background data (label 2) far from foreground data has weight 1. Foreground data (label 1) has a higher weight. Background data near the foreground (also label 2) has weight exponentially decaying down from the foreground weight. "Background-gap" data that serves to mark boundaries between foreground objects (label 3) has the highest weight.

Arguments

label: labeled dataset to turn into weights

Optional keyword arguments

scale_xy::Real: Inverse of the distance in the xy-plane, in pixels, before the background data weight is halved. Default 0.36.

scale_z::Real: Inverse of the distance in the z-plane, in pixels, before the background data weight is halved. Default 1.

metric::String: Metric to compute distance. Default taxicab (currently, no other metrics are implemented)

weight_foreground::Real: weight of foreground (1) label

weight_bkg_gap::Real: weight of background-gap (3) label

boundary_weight: weight of foreground (2) pixels adjacent to background (1 and 3) pixels. Default nothing, which uses default foreground weight.

scale_bkg_gap::Bool: whether to upweight background-gap pixels for each neuron pixel they border.

SegmentationTools.resample_img — Function

Scales an image down by using linear interpolation for the raw image, and bkg-gap priority interpoalation for the labels.

Arguments

img: image to scale

scales: array of factors to scale down (bin) by in each dimension. Must be positive.

Optional keyword arguments

dtype::String: type of data - either "raw", "label", or "weight". Default raw.

SegmentationTools.compute_mean_iou — Function

Computes the mean IOU between raw_file and a prediction_file HDF5 files.

By default, assumes a threshold of 0.5, but this can be changed with the threshold parameter.