Data Visualization API

Visualize ROIs

SegmentationTools.view_roi_3D — Function

Plots instance segmentation image img_roi, where each object is given a different color. Can also plot raw data and semantic segmentation data for comparison.

Arguments

raw: 3D raw image. If set to nothing, it will not be plotted.

predicted: 3D semantic segmentation image. If set to nothing, it will not be plotted.

img_roi: 3D instance segmentation image

Optional keyword arguments

color_brightness: minimum RGB value (out of 1) that an object will be plotted with

plot_size: size of the plot

axis: axis to project, default 3

raw_contrast: contrast of raw image, default 1

labeled_neurons: neurons that should have a specific color, as an array of arrays.

label_colors: an array of colors, one for each array in labeled_neurons

neuron_color: the color of non-labeled neurons. If not supplied, all of them will be random different colors.

overlay_intensity: intensity of ROI overlay on raw image

SegmentationTools.view_roi_2D — Function

Plots instance segmentation image img_roi, where each object is given a different color. Can also plot raw data and semantic segmentation data for comparison.

Arguments

raw: 2D raw image. If set to nothing, it will not be plotted.

predicted: 2D semantic segmentation image. If set to nothing, it will not be plotted.

img_roi: 2D instance segmentation image

Optional keyword arguments

color_brightness: minimum RGB value (out of 1) that an object will be plotted with

SegmentationTools.make_plot_grid — Function

Makes grid out of many smaller plots.

Arguments

plots: List of things to be plotted. Each item must be something that could be input to the plot function.

cols::Integer: Number of columns in array of plots to be created

plot_size: Size of resulting plot per row.

Visualize ROI Centroids

SegmentationTools.centroids_to_img — Function

Given an image of size imsize, converts centroids into an image mask of that size.

SegmentationTools.view_centroids_3D — Function

Displays the centroids of an image.

Arguments:

img: Image

centriods: Centroids of the image, to be superimposed on the image.

SegmentationTools.view_centroids_2D — Function

Displays the centroids of an image.

Arguments:

img: Image

centriods: Centroids of the image, to be superimposed on the image. They can be 3D; if they are, the first dimension will be ignored.

Visualize UNet

SegmentationTools.view_label_overlay — Function

Makes image of the raw data overlaid with a translucent label.

Arguments

img: raw data image (2D)

label: labels for the image

weight: mask of which label values to include. Pixels with a weight of 0 will be plotted in the raw, but not labeled, data; pixels with a weight of 1 will be plotted with raw data overlaid with label.

Optional keyword arguments

contrast::Real: Contrast factor for raw image. Default 1 (no adjustment)

label_intensity::Real: Intensity of label, from 0 to 1. Default 0.5.

SegmentationTools.visualize_prediction_accuracy_3D — Function

Generates an image which compares the predictions of the neural net with the label. Green = match, red = mismatch. Assumes the predictions and labels are binary 3D arrays.

Arguments

predicted: neural net predictions

actual: actual labels

weight: pixel weights; weight of 0 is ignored and not plotted.

SegmentationTools.visualize_prediction_accuracy_2D — Function

Generates an image which compares the predictions of the neural net with the label. Green = match, red = mismatch. Assumes the predictions and labels are binary 2D arrays.

Arguments

predicted: neural net predictions

actual: actual labels

weight: pixel weights; weight of 0 is ignored and not plotted.

SegmentationTools.display_predictions_3D — Function

Compares multiple different neural network predictions of the raw dataset, using an interactive slider to toggle between z-planes of the 3D dataset.

Arguments

raw: 3D raw dataset

label: labels on raw dataset. Set to nothing to avoid displaying labels (for instance, on a testing datset).

weight: weights on the labels. Set to nothing if you are not displaying labels.

predictions_array: various predictions of the raw dataset.

Optional keyword arguments

cols::Integer: maximum number of columns in the plot. Default 7.

plot_size: size of plot per row. Default (1800, 750).

axis: axis to project, default 3

display_accuracy::Bool: whether to display prediction accuracy (green for match, red for mismatch). Default true.

contrast::Real: contrast of raw image. Default 1.

SegmentationTools.display_predictions_2D — Function

Compares multiple different neural network predictions of the raw dataset, in comparison with the label and weight samples. The order of the plots is a plot of the raw data, followed by a plot of the weights, followed by plots of raw predictions and prediction vs label differential (in the order the predictions were specified in the array).

Arguments

raw: 2D raw dataset

label: labels on raw dataset. Set to nothing to avoid displaying labels (for instance, on a testing datset).

weight: weights on the labels. Set to nothing if you are not displaying labels.

predictions_array: various predictions of the raw dataset.

Optional keyword arguments

cols::Integer: maximum number of columns in the plot. Default 7.

plot_size: size of plot per row. Default (1800, 750).

display_accuracy::Bool: whether to display prediction accuracy (green for match, red for mismatch). Default true.

contrast::Real: contrast of raw image. Default 1.