Cropping API

SegmentationTools.crop_rotate! — Function

Crops and rotates a set of images.

Arguments

param_path::Dict: Dictionary containing paths to directories and a get_basename function that returns MHD file names.

param::Dict: Dictionary containing parameters, including:

crop_threshold_intensity: minimum number of standard deviations above the mean that a pixel must be for it to be categorized as part of a feature

crop_threshold_size: minimum size of a feature for it to be categorized as part of the worm

spacing_axi: Axial (z) spacing of the pixels, in um

spacing_lat: Lateral (xy) spacing of the pixels, in um

t_range: Time points to crop

ch_list: Channels to crop

dict_crop_rot_param::Dict: For each time point, the cropping parameters to use for that time point. If the cropping parameters at a time point are not found, they will be stored in the dictionary, modifying it.

save_MIP::Bool (optional): Whether to save png files. Default true

mhd_dir_key::String (optional): Key in param_path to directory to input MHD files. Default path_dir_mhd_shearcorrect

mhd_crop_dir_key::String (optional): Key in param_path to directory to output MHD files. Default path_dir_mhd_crop

mip_crop_dir_key::String (optional): Key in param_path to directory to output MIP files. Default path_dir_MIP_crop

SegmentationTools.crop_rotate — Function

Rotates and then crops an image, optionally along with its head and centroid locations.

Arguments

img: image to transform (3D)

crop_x: crop amount in x-dimension

crop_y: crop amount in y-dimension

crop_z: crop amount in z-dimension

theta: rotation amount in xy-plane

worm_centroid: centroid of the worm (to rotate around). NOT the centroids of ROIs in the worm.

Optional keyword arguments

fill: what value to put in pixels that were rotated in from outside the original image.

If kept at its default value "median", the median of the image will be used. Otherwise, it can be set to a numerical value.

degree: degree of the interpolation. Default Linear(); can set to Constant() for nearest-neighbors.

dtype: type of data in resulting image. Default Int16.

crop_pad: amount to pad the cropping in each dimension. Default 5 pixels in each dimension.

min_crop_size: minimum size of cropped image. Default [210,96,51], the UNet input size.

Outputs a new image that is the cropped and rotated version of img.

SegmentationTools.get_crop_rotate_param — Function

Generates cropping and rotation parameters from a frame by detecting the worm's location with thresholding and noise removal. The cropping parameters are designed to remove the maximum fraction of non-worm pixels.

Arguments

img: Image to crop

Optional keyword arguments

threshold_intensity: Number of standard deviations above mean for a pixel to be considered part of the worm. Default 3.

threshold_size: Number of adjacent pixels that must meet the threshold to be counted. Default 10.

SegmentationTools.uncrop_img_rois — Function

Uncrops all ROI images.

Arguments

param_path::Dict: Dictionary containing paths to files

param::Dict: Dictionary containing pipeline parameters

crop_params::Dict: Dictionary containing cropping parameters

img_size: Size of uncropped images to generate

roi_cropped_key::String (optional): Key in param_path corresponding to locations of ROI images to uncrop. Default path_dir_roi_watershed.

roi_uncropped_key::String (optional): Key in param_path corresponding to location to put uncropped ROI images. Default path_dir_roi_watershed_uncropped.

SegmentationTools.uncrop_img_roi — Function

Uncrops an ROI image.

Arguments

img_roi: ROI image to uncrop

crop_params: Dictionary of cropping parameters

img_size: Size of uncropped image

degree (optional): Interpolation method used. Default Constant() which results in nearest-neighbor interpolation

dtype (optional): Data type of uncropped image. Default Int16