Neuron Segmentation API

Semantic Segmentation

SegmentationTools.make_unet_input_h5 — Function

Makes a UNet input file. This function supports making files either for training or prediction.

Arguments

img_raw::Array: Raw image

img_label::Union{Nothing, Array}: Image label. If nothing, label and weight will not be generated.

path_h5::String: Path to HDF5 output directory.

crop (optional, default nothing): [crop_x, crop_y, crop_z], where every point with coordinates not in the given ranges is cropped out.

transpose::Bool (optional, default false): whether to transpose the x-y coordinates of the image.

weight_strategy::String (optional): method to generate weights from labels. Default and recommended is neighbors, which weights background pixels nearby foreground higher. Alternative is proportional, which will weight foreground and background constantly at a value inversely proportional to the number of pixels in those weights. The proportional weight function will ignore labels that are not 1 or 2 (including the background-gap label 3). This parameter has no effect if img_label is the empty string.

metric::String (optional): metric used to infer distance. Default (and only metric currently implemented) is taxicab. This parameter has no effect if img_label is the empty string.

scale_xy::Real (optional): Inverse of the distance in the xy-plane, in pixels, before the background data weight is halved. Default 1. This parameter has no effect if img_label is the empty string.

scale_z::Real (optional): Inverse of the distance in the z-plane, in pixels, before the background data weight is halved. Default 1. This parameter has no effect if img_label is the empty string.

weight_foreground::Real (optional, default 6): weight of foreground (1) label This parameter has no effect if img_label is the empty string.

weight_bkg_gap::Real (optional, default 10): weight of background-gap (3) label This parameter has no effect if img_label is the empty string.

boundary_weight (optional): weight of foreground (2) pixels adjacent to background (1 and 3) pixels. Default nothing, which uses default foreground weight. This parameter has no effect if img_label is the empty string.

bin_scale (optional): scale to bin image in each dimension [X,Y,Z]. Default 1,1,1.

SN_reduction_factor (optional): amount to reduce. Default 1 (no reduction)

SN_percent (optional): percentile to estimate std of image from. Default 16.

scale_bkg_gap::Bool (optional): whether to upweight background-gap pixels for each neuron pixel they border. Default false. This parameter has no effect if img_label is the empty string.

Makes UNet input files from all files in a directory. This function supports making files either for training or prediction.

Arguments

path_mhd::String: Path to raw data MHD files

path_nrrd::Union{Nothing, String}: Path to NRRD label files. If nothing, labels and weights will not be generated.

path_h5::String: Path to HDF5 output files.

crop (optional, default nothing): [crop_x, crop_y, crop_z], where every point with coordinates not in the given ranges is cropped out.

transpose::Bool (optional, default false): whether to transpose the x-y coordinates of the image.

weight_strategy::String (optional): method to generate weights from labels. Default and recommended is neighbors, which weights background pixels nearby foreground higher. Alternative is proportional, which will weight foreground and background constantly at a value inversely proportional to the number of pixels in those weights. The proportional weight function will ignore labels that are not 1 or 2 (including the background-gap label 3). This parameter has no effect if nrrd_path is the empty string.

metric::String (optional): metric used to infer distance. Default (and only metric currently implemented) is taxicab. This parameter has no effect if nrrd_path is the empty string.

scale_xy::Real (optional): Inverse of the distance in the xy-plane, in pixels, before the background data weight is halved. Default 1. This parameter has no effect if nrrd_path is the empty string.

scale_z::Real (optional): Inverse of the distance in the z-plane, in pixels, before the background data weight is halved. Default 1. This parameter has no effect if nrrd_path is the empty string.

weight_foreground::Real (optional, default 6): weight of foreground (1) label This parameter has no effect if nrrd_path is the empty string.

weight_bkg_gap::Real (optional, default 10): weight of background-gap (3) label This parameter has no effect if nrrd_path is the empty string.

boundary_weight (optional): weight of foreground (2) pixels adjacent to background (1 and 3) pixels. Default nothing, which uses default foreground weight. This parameter has no effect if nrrd_path is the empty string.

bin_scale (optional): scale to bin image in each dimension [X,Y,Z]. Default 1,1,1.

SN_reduction_factor (optional): amount to reduce. Default 1 (no reduction)

SN_percent (optional): percentile to estimate std of image from. Default 16.

scale_bkg_gap::Bool (optional): whether to upweight background-gap pixels for each neuron pixel they border. Default false. This parameter has no effect if nrrd_path is the empty string.

Generatesn HDF5 file, to be input to the UNet, out of a raw image file and a label file. Assumes 3D data.

Arguments

param_path::Dict: Dictionary containing paths to directories and a get_basename function that returns MHD file names, including:

path_dir_unet_data: Path to UNet input and output data

path_dir_mhd::String: Path to MHD files

t_range: Time points to watershed

ch_marker::Int: Marker channel

f_basename::Function: Function that returns the name of MHD files

SegmentationTools.call_unet — Function

Makes a local copy of a parameter file, modifies directories in that parameter file, then calls the UNet.

Arguments

param_path: path parameter dictionary including:

path_root_process: Root data path

path_dir_unet_data: Path to UNet input and output directory

path_unet_pred: Path to the predict.py file in pytorch-3d-unet installation

path_unet_param: Path to UNet prediction parameter file

path_unet_py_env: Path to a script that initializes the relevant environment variables for the UNet to run

Watershed-Threshold Instance Segmentation

SegmentationTools.instance_segmentation_watershed — Function

Runs watershed instance segmentation on all given frames and can output to various files (centroids, activity measurements, and image ROIs). Skips a given output method if the corresponding output directory was empty. Returns dictionary of results and a list of error frames (most likely because the worm was not in the field of view).

Arguments

param::Dict: Dictionary containing parameters, including:

seg_threshold_unet: Confidence threshold of the UNet output for a pixel to be counted as a neuron.

seg_min_neuron_size: Minimum neuron size, in voxels

seg_threshold_watershed: Confidence thresholds of the UNet output for a pixel to be counted as a neuron in each watershed step

seg_watershed_min_neuron_sizes: Minimum neuron sizes, in voxels, in each watershed step

param_path::Dict: Dictionary containing paths to directories and a get_basename function that returns MHD file names, including:

path_dir_unet_data: Path to UNet output data (input to the watershed algorithm)

path_dir_roi: Path to non-watershed ROI ouptut data

path_dir_roi_watershed: Path to watershed ROI output data

path_dir_marker_signal: Path to marker channel signal output data

path_dir_centroid: Path to centroid output data

path_dir_mhd::String: Path to MHD files

t_range: Time points to watershed

f_basename::Function: Function that returns the name of MHD files

save_centroid::Bool (optional): Whether to save centroids. Default false

save_signal::Bool (optional): Whether to save marker signal. Default false

save_roi::Bool (optional): Whether to save ROIs before watershedding. Default false

SegmentationTools.instance_segmentation_threshold — Function

Further instance segments a preliminary ROI image by thresholding UNet predictions and checking if ROIs split during thresholding.

Arguments:

img_roi: image that maps points to their current ROIs

predictions: UNet raw predictions

Optional keyword arguments

thresholds: Array of threshold values - at each value, check if an ROI was split. Default [0.7, 0.8, 0.9]

neuron_sizes: Array of neuron size values, one per threshold. Neurons that were found in a threshold that are smaller than the corresponding value are discarded and not counted for ROI split evidence. Default [5,4,4]

SegmentationTools.detect_incorrect_merges — Function

Detects incorrectly merged ROIs via thresholding. Thresholds the UNet raw output multiple times, checking if an ROI gets split into multiple, smaller ROIs at higher threshold values.

Arguments:

img_roi: Current ROIs for the image - the method checks each ROI for incorrect merging

predictions: UNet raw output (not thresholded)

thresholds: Array of threshold values - at each value, check if an ROI was split

neuron_sizes: Array of neuron size values, one per threshold. Neurons that were found in a threshold that are smaller than the corresponding value are discarded and not counted for ROI split evidence.

SegmentationTools.watershed_threshold — Function

Watersheds an ROI, taking as input its peaks (found previously via thresholding) and the UNet raw output.

Arguments:

points: Set of points in the ROI to watershed.

centroid_matches: Set of centroids in the points in question - each centroid will spawn a new ROI via watershed.

predictions: UNet raw output (not thresholded)

SegmentationTools.instance_segmentation — Function

Runs instance segmentation on a frame. Removes detected objects that are too small to be neurons.

Arguments

`predictions: UNet predictions array

Optional keyword arguments

min_neuron_size::Integer: smallest neuron size, in voxels. Default 7.

SegmentationTools.consolidate_labeled_img — Function

Converts an instance-segmentation image labeled_img to a ROI image. Ignores ROIs smaller than the minimum size min_neuron_size.

Watershed-Concave Instance Segmentation (currently not used)

SegmentationTools.instance_segment_concave — Function

Recursively segments all concave neurons in an image.

Arguments

img_roi: Image to segment

Optional keyword arguments

threshold_scale::Real: Neurons less concave than this won't be segmented. Default 0.3

num_neurons::Real: Maximum number of concave neurons per frame. Defaul 10.

zscale::Real: Scale of z-axis relative to xy plane. Default 1.

min_neuron_size::Integer: Minimum size of a neuron (in pixels). Default 10.

scale_recurse_multiply::Real: Factor to increase the concavity threshold for recursive segmentation. Default 1.5.

init_scale::Real: Amount to expand first neuron before computing location of second neuron. Default 0.7.